Fix pyflakes issues
