--- tests/test_colors.py
@@ -1,13 +1,0 @@
-import pytest
-
-from pyvista import colors
-
-
-def test_invalid_color_str_single_char():
-    with pytest.raises(ValueError):
-        colors.string_to_rgb('x')
-
-
-def test_invalid_color_str():
-    with pytest.raises(ValueError):
-        colors.string_to_rgb('not a color')
