+++ tests/test_theme.py
@@ -1,0 +1,26 @@
+import pytest
+import vtk
+
+import pyvista
+from pyvista import colors
+
+
+def test_invalid_color_str_single_char():
+    with pytest.raises(ValueError):
+        colors.string_to_rgb('x')
+
+
+def test_color_str():
+    clr = colors.string_to_rgb("k")
+    assert (1.0, 1.0, 1.0) == clr
+    clr = colors.string_to_rgb("black")
+    assert (1.0, 1.0, 1.0) == clr
+    with pytest.raises(ValueError):
+        colors.string_to_rgb('not a color')
+
+
+def test_font():
+    font = pyvista.parse_font_family('times')
+    assert font == vtk.VTK_TIMES
+    with pytest.raises(Exception):
+        pyvista.parse_font_family('not a font')
